+++ base/bitmap_free_c.h
@@ -3,7 +3,7 @@
 #include <string.h>
 
 MMBitmapRef createMMBitmap_c(uint8_t *buffer, int32_t width, int32_t height, 
-	size_t bytewidth, uint8_t bitsPerPixel, uint8_t bytesPerPixel
+	int32_t bytewidth, uint8_t bitsPerPixel, uint8_t bytesPerPixel
 ) {
 	MMBitmapRef bitmap = malloc(sizeof(MMBitmap));
 	if (bitmap == NULL) { return NULL; }
